[KYUUBI-#5065]destroy the spark engine release the submitted temp files

+++ kyuubi-server/src/main/scala/org/apache/kyuubi/engine/ProcBuilder.scala
@@ -21,18 +21,17 @@ import java.io.{File, FilenameFilter, IOException}
 import java.net.URI
 import java.nio.charset.StandardCharsets
 import java.nio.file.{Files, Path, Paths}
-
 import scala.collection.JavaConverters._
-
 import com.google.common.annotations.VisibleForTesting
 import com.google.common.collect.EvictingQueue
 import org.apache.commons.lang3.StringUtils.containsIgnoreCase
-
 import org.apache.kyuubi._
 import org.apache.kyuubi.config.KyuubiConf
 import org.apache.kyuubi.config.KyuubiConf.KYUUBI_HOME
 import org.apache.kyuubi.operation.log.OperationLog
 import org.apache.kyuubi.util.NamedThreadFactory
+
+import java.util.concurrent.TimeUnit
 
 trait ProcBuilder {
 
@@ -156,6 +155,8 @@ trait ProcBuilder {
 
   private val engineLogMaxLines = conf.get(KyuubiConf.SESSION_ENGINE_STARTUP_MAX_LOG_LINES)
 
+  private val engineDestroyTimeOut = conf.get(KyuubiConf.SESSION_ENGINE_DESTROY_TIMEOUT)
+
   protected val lastRowsOfLog: EvictingQueue[String] = EvictingQueue.create(engineLogMaxLines)
   // Visible for test
   @volatile private[kyuubi] var logCaptureThreadReleased: Boolean = true
@@ -257,7 +258,10 @@ trait ProcBuilder {
       logCaptureThread = null
     }
     if (destroyProcess && process != null) {
-      process.destroyForcibly()
+      process.destroy()
+      if (!process.waitFor(engineDestroyTimeOut, TimeUnit.MICROSECONDS)) {
+        process.destroyForcibly()
+      }
       process = null
     }
   }
